refactor(cli): standardize command directory structure
Convert all CLI commands to consistent directory-style pattern:
- Move auth subcommands (logout, status, setup-token) to separate files
- Create info command directory
- Convert single-file commands to directories (compose, init, usage, onboard, setup-claude)
- Split cook command into separate files for main action and subcommands
- Clean up index.ts to only import and register commands
